requirements.txt
ultralytics==8.*
torch
torchvision
opencv-python-headless
numpy
pandas
tqdm
imageio
streamlit
pyyaml